refactor(reusableComponents): review changes
