--- Components/BlockReusable/README.md
@@ -1,3 +1,0 @@
-# Block Reusable
-
-Set up reusable posts with a component or combination of components. Select one or more reusable posts to show their components.
