+++ Components/ReusableComponents/README.md
@@ -1,0 +1,3 @@
+# Reusable Components
+
+Set up reusable posts with a component or combination of components. Select one or more reusable components.
