+++ Components/ReusableComponents/functions.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Flynt\Components\BlockReusable;
+namespace Flynt\Components\ReusableComponents;
 
 use Timber\Timber;
 
-add_filter('Flynt/addComponentData?name=BlockReusable', function ($data) {
+add_filter('Flynt/addComponentData?name=ReusableComponents', function ($data) {
     if (!empty($data['ids'])) {
         $data['posts'] = Timber::get_posts([
             'post_type' => 'reusable-component',
@@ -18,8 +18,8 @@ add_filter('Flynt/addComponentData?name=BlockReusable', function ($data) {
 function getACFLayout()
 {
     return [
-        'name' => 'blockReusable',
-        'label' => 'Block: Reusable',
+        'name' => 'reusableComponents',
+        'label' => 'Reusable: Components',
         'sub_fields' => [
             [
                 'label' => __('Components', 'flynt'),
@@ -33,7 +33,7 @@ function getACFLayout()
                 'ui' => 1,
                 'required' => 1,
                 'return_format' => 'id',
-                'instructions' => 'Select reusable posts to show their components. You can manage reusable components <a href="/wp/wp-admin/edit.php?post_type=reusable-component">here</a>.'
+                'instructions' => 'Select reusable components. Manage <a href="/wp/wp-admin/edit.php?post_type=reusable-component">reusable components</a>.'
             ],
         ]
     ];
