+++ inc/fieldGroups/pageComponents.php
@@ -1,7 +1,7 @@
 <?php
 
 use ACFComposer\ACFComposer;
-use Flynt\SetComponents;
+use Flynt\Components;
 
 add_action('Flynt/afterRegisterComponents', function () {
     ACFComposer::registerFieldGroup([
@@ -15,8 +15,17 @@ add_action('Flynt/afterRegisterComponents', function () {
                 'type' => 'flexible_content',
                 'button_label' => __('Add Component', 'flynt'),
                 'layouts' => [
-                    SetComponents\getPageComponents()
-                ]
+                    Components\BlockCollapse\getACFLayout(),
+                    Components\BlockImage\getACFLayout(),
+                    Components\BlockImageText\getACFLayout(),
+                    Components\BlockVideoOembed\getACFLayout(),
+                    Components\BlockWysiwyg\getACFLayout(),
+                    Components\GridImageText\getACFLayout(),
+                    Components\GridPostsLatest\getACFLayout(),
+                    Components\ListComponents\getACFLayout(),
+                    Components\SliderImages\getACFLayout(),
+                    Components\ReusableComponents\getACFLayout(),
+                ],
             ]
         ],
         'location' => [
@@ -30,8 +39,8 @@ add_action('Flynt/afterRegisterComponents', function () {
                     'param' => 'post_type',
                     'operator' => '!=',
                     'value' => 'reusable-component'
-                ]
-            ]
-        ]
+                ],
+            ],
+        ],
     ]);
 });
